Changing the function used to move
(old and test functions were commented)
I moved the movement methods from the previous Rigidbody2D.velocity.
Previous move method would move properly but would get stuck to walls...
I then went around the Internet and found Rigidbody2D.AddForce. but for
ground movement it doesn't work at all.
Next try used transform.Translate and this time it looks like it does
work, though I'm not sure how it works...
Basically though, it looks like it moves an object... and it can be
turned based on object rotations with the Space variable... looks
interesting...
...
I'll try to explain better:
It receives a Vector3 (I used a Vector2 instead, because 2D, and it
works) that's for the movement, and then (optionally) receives a
variable that shows what that movement's going to be relative about.
Looks like it defaults to Space.Self which uses the object's own
rotation but it can also use Space.World which's based on the world
coordinates.
Next thing I plan to do is actually doing a function for wall
interacting... whatever that is...